Update 타운 복귀 bull queue 적용
